requirements.txt
torch>=1.9.0
transformers>=4.15.0
datasets>=2.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
streamlit>=1.8.0
nltk>=3.6.0
pyyaml>=6.0
tqdm>=4.62.0
tensorboard>=2.8.0